Refactored for Svelte 4 framework
