+++ index.html
@@ -1,10 +1,15 @@
+<!-- 
+  Main entry point for the plugin UI.
+  Svelte is controlling the app, so one div in body that
+  will be the target for the Svelte UI renderer.
+-->
+
 <!DOCTYPE html>
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>VCS</title>
-    <script type="text/javascript" src="./nel-vcs-24/js/patch.js"></script>
   </head>
 
   <body>
