+++ public/manifest.json
@@ -1,14 +1,14 @@
 {
   "window": {
-    "width": 753,
-    "height": 373
+    "width": 800,
+    "height": 494
   },
   "parameters": [
-    { "paramId": "size",    "name": "Size",     "min": 0.0, "max": 1.0, "defaultValue": 0.5 },
-    { "paramId": "decay",   "name": "Decay",    "min": 0.0, "max": 1.0, "defaultValue": 0.5 },
-    { "paramId": "mod",     "name": "Mod",      "min": 0.0, "max": 1.0, "defaultValue": 0.5 },
-    { "paramId": "mix",     "name": "Mix",      "min": 0.0, "max": 1.0, "defaultValue": 0.5 },
-    { "paramId": "circleID","name": "Circle ID",  "min": 0, "max": 1, "defaultValue": 0.5 },
-    { "paramId" : "nodeValue" , "name" : "Node Value", "min": 0.0, "max": 1.0, "defaultValue": 0.5 }
+    { "paramId": "size", "name": "Size", "min": 0.0, "max": 1.0, "defaultValue": 0.5 },
+    { "paramId": "decay", "name": "Decay", "min": 0.0, "max": 1.0, "defaultValue": 0.5 },
+    { "paramId": "mod", "name": "Mod", "min": 0.0, "max": 1.0, "defaultValue": 0.5 },
+    { "paramId": "mix", "name": "Mix", "min": 0.0, "max": 1.0, "defaultValue": 0.5 },
+    { "paramId": "circleID", "name": "Circle ID", "min": 0, "max": 1, "defaultValue": 0.5 },
+    { "paramId": "nodeValue", "name": "Node Value", "min": 0.0, "max": 1.0, "defaultValue": 0.5 }
   ]
 }
