+++ src/lib/CablesUI.svelte
@@ -6,12 +6,14 @@
 
   let pathPatch: string = `/${patch}/patch.js`;
 
+  // move Cables assets folder up to /public or they don't get found
+
   const initializeCables = () => {
     CablesPatch.set(
       new CABLES.Patch({
         patch: CABLES.exportedPatch,
-        prefixAssetPath: ``,
-        assetPath: "",
+        prefixAssetPath: "",
+        assetPath: "../assets/",
         jsPath: pathPatch,
         glCanvasId: `cables_${patch}`,
         glCanvasResizeToWindow: true,
@@ -27,14 +29,14 @@
 
   function showError(errId: any, errMsg: any) {
     // handle critical errors here if needed
-    console.error("CRITICAL ERROR thrown from UI");
+    console.error("BLOODY ERROR from UI");
   }
 
   // Cables patch initialized, set up interop bindings
   // prepare for connecting patch vars to Elementary native
 
   function patchInitialized(patch: any) {
-    console.log("patch initialized", CABLES);
+    console.log("patch initialized", patch);
 
     // get  all the variables from the Cables patch
     const cablesVariables = patch.getVars();
@@ -56,7 +58,7 @@
         paramId !== "interpolatingPreset" &&
         paramId !== "paramsAverage"
       ) {
-        value.on("change", function (newValue) {
+        value.on("change", function (newValue: number) {
           requestParamValueUpdate(paramId, newValue);
         });
       }
@@ -76,7 +78,7 @@
     // use the average as a change trigger
 
     if (paramsAverage) {
-      paramsAverage.on("change", function (val) {
+      paramsAverage.on("change", function (newValue: number) {
         const values = interpolatingPreset.getValue();
         if (values && values.length >= 4) {
           const params = ["decay", "mix", "mod", "size"];
@@ -93,56 +95,24 @@
     // it needs to be divided by 2 because the preset index is 2->72
     // but actually there are only 36 presets
     if (pickedID) {
-      pickedID.on("change", function (id) {
-        const element = getConsoleDiv();
-        clearInterval(currentTimer);
-        element.style.opacity = 1;
-        currentTimer = setInterval(() => {
-          element.style.opacity = 0;
-        }, 3000);
+      pickedID.on("change", function (id: number) {
         // store it in a global variable
         currentID = id / 2;
-
         // log to the console inside the patch
         const stem = "‣ ";
         patch.setVariable("topLeftText", stem + "Node » " + currentID);
       });
     }
-
-    // readout is the value generated by dragging the mouse over the circle
-    // requestParamValueUpdate is called here with the circleID and the readout value
-    if (readout) {
-      readout.on("change", function (newValue) {
-        const element = getConsoleDiv();
-        element.classList.add("fadeIn");
-        requestParamValueUpdate("circleID", currentID / 36);
-        requestParamValueUpdate("nodeValue", newValue);
-      });
-    }
   }
 
-  function patchFinishedLoading(patch) {
+  function patchFinishedLoading() {
     // The patch is ready now, all assets have been loaded
     // not used so far
   }
 
-  document.addEventListener("CABLES.jsLoaded", function (event) {
-    CABLES.patch = new CABLES.Patch({
-      patch: CABLES.exportedPatch,
-      prefixAssetPath: "",
-      assetPath: "assets/",
-      jsPath: "",
-      glCanvasId: "glcanvas",
-      glCanvasResizeToWindow: true,
-      onError: showError,
-      onPatchLoaded: patchInitialized,
-      onFinishedLoading: patchFinishedLoading,
-      canvas: { alpha: true, premultipliedAlpha: true }, // make canvas transparent
-    });
-  });
-
   // Interop bindings
-  function requestParamValueUpdate(paramId, value) {
+  function requestParamValueUpdate(paramId: string, value: number) {
+    let globalThis = window as any;
     if (typeof globalThis.__postNativeMessage__ === "function") {
       globalThis.__postNativeMessage__("setParameterValue", {
         paramId,
@@ -150,7 +120,18 @@
       });
     }
   }
-  onMount(() => {
+  onMount(async () => {
+    try {
+      await new Promise((resolve, reject) => {
+        const script = document.createElement("script");
+        script.src = "/nel-vcs-24/js/patch.js";
+        script.onload = resolve;
+        script.onerror = reject;
+        document.head.appendChild(script);
+      });
+    } catch (error) {
+      console.error("Error loading Cables Patch", error);
+    }
     initializeCables();
   });
 </script>
