--- src/lib/Counter.svelte
@@ -1,10 +1,0 @@
-<script lang="ts">
-  let count: number = 0;
-  const increment = () => {
-    count += 1;
-  };
-</script>
-
-<button on:click={increment}>
-  count is {count}
-</button>
