+++ src/main.ts
@@ -2,7 +2,7 @@ import "./app.css";
 import App from "./App.svelte";
 
 const app = new App({
-  target: document.getElementById("app"),
+  target: document.getElementById("app") || document.body,
 });
 
 export default app;
